python/requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
azure-core==1.32.0
azure-identity==1.19.0
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
jmespath==1.0.1
msal==1.31.0
msal-extensions==1.2.0
openai==1.54.3
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
